Margaret of Austria, Duchess of Savoy (1480-1530)
Margaret of Austria, Duchess of Savoy (1480-1530) is the daughter of Maximilian I, Holy Roman Emperor & his wife Mary of Burgundy. The late wife of John, Prince of Asturias and Girona (b. 1478, m. 1497 & d. 1497) and then the late wife of Philibert II, Duke of Savoy (b. 1480, m. 1501 & d. 1504).
This is a genetic file for her, which can be pasted into CK3's ruler designer.